backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
pytest-cov==4.1.0

# DocuSign SDK
docusign-esign==3.26.0

# Utilities
email-validator==2.1.0.post1
python-magic==0.4.27

# Phase 2: Email & Notifications
aiosmtplib==3.0.1
jinja2==3.1.3

# Phase 2: Microsoft Teams Integration
msgraph-sdk==1.1.0
azure-identity==1.15.0

# Phase 2: File handling
aiofiles==23.2.1
